GSdx: Restore Unscale PointLine as a Disable option and combine it with GS Memory clear and name it Disable Safe Features.
Idea is to disable multiple features in 1 option to avoid gui clutter.

This partially reverts PR #2710 restoring Unscale Point and Line but as a disable option.
It caused black surfaces to appear when upscaling  on Xenosaga series so it looks
like the option isn't fully safe yet to be always enabled.

This option is also now Merged with GS memory clear as well so it's effectively 2 options
in to one that cover some rare corner cases where these two features aren't supported the
option can be enabled.